Improved docs of all modules under sorting...
I want to generate automatically html files representing the docs of
this project from the docstrings using the tool pdoc (which works
fine), so I am converting the docstrings of every module conforming its
rules (markdown, etc). I have also making ands simply a package and all
imports will be relative (probably). Improved tests for sorting
algorithms. Added a new switch or swap or replace function for BST,
which should be able to swap any to arbitrary nodes in a bst… preparing
to write the delete function of a RBT…