Fixed a bug in the MinMaxHeap class
The deletion method implemented in the previous Heap class was not
valid for the MinMaxHeap but only for MinHeap and MaxHeap. So I created
another level between the base heap and the min heap and the max heap,
so that min-max heap does not have to inherit code that doesn’t work.